+++ src/support/serde.rs
@@ -1,29 +1,25 @@
 #![cfg(feature = "serde")]
-use crate::{nbytes, nlimbs, Uint};
+use crate::{nbytes, Uint};
 use core::fmt::{Formatter, Result as FmtResult};
 use serde::{
     de::{Error, Unexpected, Visitor},
     Deserialize, Deserializer, Serialize, Serializer,
 };
-use std::str;
+use std::{fmt::Write, str};
 
 /// Serialize a [`Uint`] value.
 ///
 /// For human readable formats a `0x` prefixed lower case hex string is used.
 /// For binary formats a byte array is used. Leading zeros are included.
-impl<const BITS: usize> Serialize for Uint<BITS>
-where
-    [(); nlimbs(BITS)]:,
-    [(); nbytes(BITS)]:,
-{
+impl<const BITS: usize, const LIMBS: usize> Serialize for Uint<BITS, LIMBS> {
     fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        let bytes = self.to_be_bytes();
+        let bytes = self.to_be_bytes_vec();
         if serializer.is_human_readable() {
             // TODO: Without allocations
             let mut result = String::with_capacity(2 * Self::BYTES + 2);
             result.push_str("0x");
-            for byte in bytes.iter() {
-                result.push_str(&format!("{:02x}", byte));
+            for byte in bytes {
+                write!(result, "{:02x}", byte).unwrap();
             }
             serializer.serialize_str(&result)
         } else {
@@ -37,11 +33,7 @@ where
 /// Hex strings can be upper/lower/mixed case, have an optional `0x` prefix, and
 /// can be any length. They are interpreted big-endian.
 // TODO: Document and test the range of valid inputs.
-impl<'de, const BITS: usize> Deserialize<'de> for Uint<BITS>
-where
-    [(); nlimbs(BITS)]:,
-    [(); nbytes(BITS)]:,
-{
+impl<'de, const BITS: usize, const LIMBS: usize> Deserialize<'de> for Uint<BITS, LIMBS> {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
         if deserializer.is_human_readable() {
             deserializer.deserialize_str(StrVisitor)
@@ -52,17 +44,10 @@ where
 }
 
 /// Serde Visitor for human readable formats
-struct StrVisitor<const BITS: usize>
-where
-    [(); nlimbs(BITS)]:,
-    [(); nbytes(BITS)]:;
+struct StrVisitor<const BITS: usize, const LIMBS: usize>;
 
-impl<'de, const BITS: usize> Visitor<'de> for StrVisitor<BITS>
-where
-    [(); nlimbs(BITS)]:,
-    [(); nbytes(BITS)]:,
-{
-    type Value = Uint<BITS>;
+impl<'de, const BITS: usize, const LIMBS: usize> Visitor<'de> for StrVisitor<BITS, LIMBS> {
+    type Value = Uint<BITS, LIMBS>;
 
     fn expecting(&self, formatter: &mut Formatter) -> FmtResult {
         write!(formatter, "a {} byte hex string", nbytes(BITS))
@@ -73,7 +58,7 @@ where
         E: Error,
     {
         let value = trim_hex_prefix(value);
-        let mut limbs = [0; nlimbs(BITS)];
+        let mut limbs = [0; LIMBS];
         for (i, chunk) in value.as_bytes().rchunks(16).enumerate() {
             let chunk = str::from_utf8(chunk)
                 .map_err(|_| Error::invalid_value(Unexpected::Str(value), &self))?;
@@ -82,12 +67,12 @@ where
             if limb == 0 {
                 continue;
             }
-            if i >= nlimbs(BITS) {
+            if i >= LIMBS {
                 return Err(Error::invalid_value(Unexpected::Str(value), &self));
             }
             limbs[i] = limb;
         }
-        if BITS > 0 && limbs[nlimbs(BITS) - 1] > Self::Value::MASK {
+        if BITS > 0 && limbs[LIMBS - 1] > Self::Value::MASK {
             return Err(Error::invalid_value(Unexpected::Str(value), &self));
         }
         Ok(Uint::from_limbs(limbs))
@@ -95,17 +80,10 @@ where
 }
 
 /// Serde Visitor for non-human readable formats
-struct ByteVisitor<const BITS: usize>
-where
-    [(); nlimbs(BITS)]:,
-    [(); nbytes(BITS)]:;
+struct ByteVisitor<const BITS: usize, const LIMBS: usize>;
 
-impl<'de, const BITS: usize> Visitor<'de> for ByteVisitor<BITS>
-where
-    [(); nlimbs(BITS)]:,
-    [(); nbytes(BITS)]:,
-{
-    type Value = Uint<BITS>;
+impl<'de, const BITS: usize, const LIMBS: usize> Visitor<'de> for ByteVisitor<BITS, LIMBS> {
+    type Value = Uint<BITS, LIMBS>;
 
     fn expecting(&self, formatter: &mut Formatter) -> FmtResult {
         write!(formatter, "{BITS} bits of binary data in big endian order")
@@ -118,7 +96,7 @@ where
         if value.len() != nbytes(BITS) {
             return Err(E::invalid_length(value.len(), &self));
         }
-        Uint::try_from_be_bytes(value).ok_or_else(|| {
+        Uint::try_from_be_slice(value).ok_or_else(|| {
             E::invalid_value(
                 Unexpected::Other(&format!("Value to large for Uint<{}", BITS)),
                 &self,
@@ -139,13 +117,14 @@ fn trim_hex_prefix(str: &str) -> &str {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::const_for;
+    use crate::{const_for, nlimbs};
     use proptest::proptest;
 
     #[test]
     fn test_serde_human_readable() {
         const_for!(BITS in SIZES {
-            proptest!(|(value: Uint<BITS>)| {
+            const LIMBS: usize = nlimbs(BITS);
+            proptest!(|(value: Uint<BITS, LIMBS>)| {
                 let serialized = serde_json::to_string(&value).unwrap();
                 let deserialized = serde_json::from_str(&serialized).unwrap();
                 assert_eq!(value, deserialized);
@@ -156,7 +135,8 @@ mod tests {
     #[test]
     fn test_serde_machine_readable() {
         const_for!(BITS in SIZES {
-            proptest!(|(value: Uint<BITS>)| {
+            const LIMBS: usize = nlimbs(BITS);
+            proptest!(|(value: Uint<BITS, LIMBS>)| {
                 let serialized = bincode::serialize(&value).unwrap();
                 let deserialized = bincode::deserialize(&serialized[..]).unwrap();
                 assert_eq!(value, deserialized);